srcs/inventory-app/requirements.txt
blinker==1.9.0
click==8.3.3
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.8
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-SQLAlchemy
requests
